get dnsmasq process: lookup in proc FS
Golang`s `os.FindProcess()` always returns success on UNIX machines -
even when the PID doesn't exist.  Hence, add a lookup in the proc FS
to check if the given exists or not.  That's okay since dnsmane is
targetting Linux only.

Also rename the method to `getProcess()` and add some comments.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>