Add AppArmor instruction in the Podman readme
Users can experience issues if Apparmor has the dnsmasq profile in
enforce mode. This commit adds instruction on how to setup a local
override for the dnsmasq profile to give dnsmasq the necessary file
access permissions.

The profile has been tested with the commands on the tutorial and the
commands worked without AppArmor reporting blocked file access
operations.

Related issues:

  https://github.com/containers/podman/issues/9364
  https://github.com/containers/dnsname/issues/33

Signed-off-by: Michele Sorcinelli <michelesr@autistici.org>