Add dns search domain to cni result
Local host name lookup is slow and can time out because
the search domain is not added to resolv.conf.

We have to add the configured domain name to the dns search domain
cni result so podman can read it back and add it to resolv.conf,
similar to how it is done for the nameservers.

Signed-off-by: Paul Holzinger <paul.holzinger@web.de>