only delete dnsmasq.conf in cleanup
This fixes a race condition with deleting the last network and adding a
new one.

Previously the entire directory would be removed, and if this happened
in between an `add` call's stat(configDir) and flock(configDir/lock) the
stat() would see that the dir is there but then the flock() would error
trying to open(configDir/lock).

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>